Revert "Sentiment pretraining"
This reverts commit a93d5695a6a3f729453366da6a796ffdd08ae50a.
